Move PaymentService into it's own class.
