+++ components/PayPalButton.vue
@@ -14,6 +14,10 @@
 </template>
 
 <script>
+/*!
+ * Copyright (c) 2018-2019 Digital Bazaar, Inc. All rights reserved.
+ */
+'use strict';
 
 /**
  * PayPalButton.vue is a simple wrapper around paypal's smart buttons.
