+++ lib/Payment.js
@@ -1,6 +1,7 @@
 /*!
  * Copyright (c) 2018-2019 Digital Bazaar, Inc. All rights reserved.
  */
+'use strict';
 
 /**
  * Payment is class that tracks payments from various providers.
